Imrprove code design
Avoid the Back Office being dependent on the Runtime

Refactor for #1753
